show a popup in full screen or just over thacars with a timer which is couting down untill new free generation are available. also give three buttons one is close, one is access the community and one is unlock unlimited.
unlock unlimited redirects to "https://creatorpreneurclub.superprofile.bio/india/?referralCode=". and instet the referral code fond or stored in cache here.

also make the life of this cache to 30 days and make it changable if a new value is found wothin that time limit and then reset the time of statage when new value is found